extract dsl.rb
